requirements.txt
asgiref==3.6.0
Django==4.1.7
django-recaptcha==3.0.0
load-dotenv==0.1.0
psycopg2==2.9.5
python-dotenv==0.21.1
sqlparse==0.4.3
